toolkit/devtools/qrcode/tests/mochitest/test_decode.html

<!DOCTYPE HTML>
<html>
<!--
Test decoding a simple message
-->
<head>
  <meta charset="utf-8">
  <title>Test decoding a simple message</title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8">
window.onload = function() {
  const { utils: Cu } = Components;
  const { devtools } = Cu.import("resource://gre/modules/devtools/Loader.jsm", {});
  const { require } = devtools;
  const { Task } = Cu.import("resource://gre/modules/Task.jsm", {});
  const { Promise: promise } =
    Cu.import("resource://gre/modules/Promise.jsm", {});

  const QR = require("devtools/toolkit/qrcode/index");

  SimpleTest.waitForExplicitFinish();

  const testImage =
    "data:image/gif;base64,R0lGODdhOgA6AIAAAAAAAP///ywAAAAAOgA6AAAC" +
    "/4yPqcvtD6OctNqLs968+w+G4gKU5nkaKKquLuW+QVy2tAkDTj3rfQts8CRDko" +
    "+HPPoYRUgy9YsyldDm44mLWhHYZM6W7WaDqyCRGkZDySxpRGw2sqvLt1q5w/fo" +
    "XyE6vnUQOJUHBlinMGh046V1F5PDqNcoqcgBOWKBKbK2N+aY+Ih49VkmqMcl2l" +
    "dkhZUK1umE6jZXJ2ZJaujZaRqH4bpb2uZrJxvIt4Ebe9qoYYrJOsw8apz2bCut" +
    "m9kqDcw52uuImyr5Oh1KXH1jrn2anuunywtODU/o2c6teceW39ZcLFg/fNMo1b" +
    "t3jVw2dwTPwJq1KYG3gAklCgu37yGxeScYKyiCc+7DR34hPVQiuQ7UhJMagyEb" +
    "lymmzJk0a9q8iTOnzp0NCgAAOw==";

  Task.spawn(function() {
    let result = yield QR.decodeFromURI(testImage);
    is(result, "HELLO", "Decoded data URI result matches");
    let canvas = yield drawToCanvas(testImage);
    result = QR.decodeFromCanvas(canvas);
    is(result, "HELLO", "Decoded canvas result matches");
  }).then(SimpleTest.finish, ok.bind(null, false));

  function drawToCanvas(src) {
    let deferred = promise.defer();
    let canvas = document.createElement("canvas");
    let context = canvas.getContext("2d");
    let image = new Image();

    image.onload = () => {
      canvas.width = image.width;
      canvas.height = image.height;
      context.drawImage(image, 0, 0);
      deferred.resolve(canvas);
    };
    image.src = src;

    return deferred.promise;
  }
};
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>
</html>
